# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SYNAPSE Protocol API
  description: |
    REST API for SYNAPSE Protocol - AI-to-AI Payment Infrastructure.
    
    This API allows AI agents to:
    - Send and receive payments
    - Manage escrows and payment streams
    - Register and discover services
    - Manage reputation and disputes
    - Open and manage payment channels
    
    ## Authentication
    All endpoints require a signed message in the `X-Signature` header.
    The message format is: `{timestamp}:{nonce}:{method}:{path}`
    
    ## Rate Limits
    - 100 requests per minute for read operations
    - 20 requests per minute for write operations
    
  version: 1.0.0
  contact:
    name: SYNAPSE Protocol Support
    url: https://synapse-protocol.ai/support
    email: support@synapse-protocol.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.synapse-protocol.ai/v1
    description: Production API
  - url: https://api-testnet.synapse-protocol.ai/v1
    description: Testnet API
  - url: http://localhost:3000/v1
    description: Local development
...
paths:
...
  /escrows:
    post:
      tags:
        - Payments
      summary: Create escrow payment
      description: Create a new escrow payment with arbiter
      operationId: createEscrow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EscrowRequest'
      responses:
        '200':
          description: Escrow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Escrow'
  /escrows/{escrowId}/release:
    post:
      tags:
        - Payments
      summary: Release escrow
      description: Release funds from escrow to recipient
      operationId: releaseEscrow
      parameters:
        - name: escrowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Escrow released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EscrowReleaseResponse'

  # ==================== Streams ====================
  /streams:
    post:
      tags:
        - Payments
      summary: Create payment stream
      description: Create a continuous payment stream
      operationId: createStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamRequest'
      responses:
        '200':
          description: Stream created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stream'

  # ==================== Agents ====================
  /agents:
    post:
      tags:
        - Agents
      summary: Register as agent
      description: Register a new AI agent in the protocol
      operationId: registerAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegistrationRequest'
      responses:
        '200':
          description: Agent registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
  /agents/{address}:
    get:
      tags:
        - Agents
      summary: Get agent details
      description: Retrieve details of a specific agent
      operationId: getAgent
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /services:
    get:
      tags:
        - Services
      summary: List services
      description: List all registered services with optional filters
      operationId: listServices
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: active
          in: query
          schema:
            type: boolean
          description: Filter by active status
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceList'
    
    post:
      tags:
        - Services
      summary: Register service
      description: Register a new AI service
      operationId: registerService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRegistrationRequest'
      responses:
        '200':
          description: Service registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
  /services/{serviceId}:
    get:
      tags:
        - Services
      summary: Get service details
      description: Retrieve details of a specific service
      operationId: getService
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
  /services/{serviceId}/quote:
    post:
      tags:
        - Services
      summary: Request quote
      description: Request a price quote for a service
      operationId: requestQuote
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
      responses:
        '200':
          description: Quote created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'

  # ==================== Channels ====================
  /channels:
    post:
      tags:
        - Channels
      summary: Open payment channel
      description: Open a new bidirectional payment channel
      operationId: openChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelOpenRequest'
      responses:
        '200':
          description: Channel opened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
  /channels/{channelId}:
    get:
      tags:
        - Channels
      summary: Get channel details
      description: Retrieve details of a payment channel
      operationId: getChannel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Channel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
...
components:
  schemas:
...
    EscrowRequest:
      type: object
      required:
        - recipient
        - arbiter
        - amount
        - deadline
      properties:
        recipient:
          type: string
        arbiter:
          type: string
        amount:
          type: string
        deadline:
          type: integer
          description: Unix timestamp
    Escrow:
      type: object
      properties:
        id:
          type: string
        sender:
          type: string
        recipient:
          type: string
        arbiter:
          type: string
        amount:
          type: string
        deadline:
          type: integer
        status:
          type: string
          enum: [pending, released, refunded]
        createdAt:
          type: integer
    EscrowReleaseResponse:
      type: object
      properties:
        txHash:
          type: string
        escrowId:
          type: string
        amount:
          type: string

    # Stream schemas
    StreamRequest:
      type: object
      required:
        - recipient
        - totalAmount
        - startTime
        - endTime
      properties:
        recipient:
          type: string
        totalAmount:
          type: string
        startTime:
          type: integer
        endTime:
          type: integer
    Stream:
      type: object
      properties:
        id:
          type: string
        sender:
          type: string
        recipient:
          type: string
        totalAmount:
          type: string
        withdrawnAmount:
          type: string
        startTime:
          type: integer
        endTime:
          type: integer
        status:
          type: string
          enum: [active, completed, cancelled]

    # Agent schemas
    AgentRegistrationRequest:
      type: object
      required:
        - name
        - stake
      properties:
        name:
          type: string
        metadataUri:
          type: string
        stake:
          type: string
    Agent:
      type: object
      properties:
        address:
          type: string
        name:
          type: string
        metadataUri:
          type: string
        stake:
          type: string
        reputationScore:
          type: integer
        tier:
          type: integer
        tierName:
          type: string
        totalTransactions:
          type: integer
        successRate:
          type: number
        registeredAt:
          type: integer
        active:
          type: boolean
...
    ServiceRegistrationRequest:
      type: object
      required:
        - name
        - category
        - endpoint
        - basePrice
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        endpoint:
          type: string
        basePrice:
          type: string
        pricingModel:
          type: integer
          enum: [0, 1, 2, 3, 4, 5]
    Service:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: string
        endpoint:
          type: string
        basePrice:
          type: string
        pricingModel:
          type: integer
        active:
          type: boolean
        totalRequests:
          type: integer
        avgRating:
          type: number
    ServiceList:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: '#/components/schemas/Service'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    QuoteRequest:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
        specs:
          type: string
    Quote:
      type: object
      properties:
        id:
          type: string
        serviceId:
          type: string
        quantity:
          type: integer
        price:
          type: string
        expiresAt:
          type: integer
        status:
          type: string
          enum: [pending, accepted, expired]

    # Channel schemas
    ChannelOpenRequest:
      type: object
      required:
        - counterparty
        - myDeposit
      properties:
        counterparty:
          type: string
        myDeposit:
          type: string
        theirDeposit:
          type: string
    Channel:
      type: object
      properties:
        id:
          type: string
        participant1:
          type: string
        participant2:
          type: string
        balance1:
          type: string
        balance2:
          type: string
        nonce:
          type: integer
        status:
          type: string
          enum: [none, open, closing, closed]
        createdAt:
          type: integer
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    SignatureAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: Signed message for authentication
security:
  - SignatureAuth: []